Blog Update
Update for authorizations, and have likes and dislikes for blogs

+++ lvlgg_backend/comment/views.py
@@ -2,6 +2,7 @@ from django.http import JsonResponse, Http404
 from account.models import Client
 from blog.models import Blog
 from django.shortcuts import get_object_or_404
+from django.contrib.auth import authenticate
 from django.views.decorators.http import require_http_methods
 from django.views import View
 import json
@@ -10,6 +11,10 @@ from .models import Comment
 class Comments(View):
 
     def post(self, request):
+
+        if not request.user.is_authenticated:
+            return JsonResponse({'error': 'Not Logged In'}, status=404)
+        
         data = json.loads(request.body.decode('utf-8')) 
         content = data.get('content')
         author_id = data.get('author')
@@ -41,6 +46,10 @@ class Comments(View):
     
 
     def delete(self, request, pk):
+
+        if not request.user.is_authenticated:
+            return JsonResponse({'error': 'Not Logged In'}, status=404)
+        
         try:
             comment = get_object_or_404(Comment, pk=pk)
             comment.delete()
@@ -52,6 +61,10 @@ class Comments(View):
     
     # Assuming you want to update comments as well
     def put(self, request, pk):
+
+        if not request.user.is_authenticated:
+            return JsonResponse({'error': 'Not Logged In'}, status=404)
+        
         try:
             comment = get_object_or_404(Comment, pk=pk)
             data = json.loads(request.body.decode('utf-8'))
